refactor(turns): decompose turn_state_engine.rs into a directory module
`turn_state_engine.rs` had grown to ~4,310 lines (well past the 1,500/3,000
architecture thresholds; it carried an `arch-exempt: large_file`). Split it into
`turn_state_engine/{mod,limits,run_record,idempotency,snapshot,admission,
spawn_tree,transitions}.rs`:

- mod.rs (1,397) — TurnStateEngine/Inner/RunRecord/key struct defs, RunStatusCell,
  small shared free fns, impl TurnStateEngine, and the
  TurnEventProjectionSource/LoopCheckpointStore/TurnStateStore trait impls.
- transitions.rs (1,451) — the Inner run-transition family + impl
  TurnRunTransitionPort + lease recovery + AppliedLoopTransition.
- spawn_tree.rs (515) — impl TurnSpawnTreeStateStore.
- idempotency.rs (340) — the idempotency ledger methods + record builders.
- snapshot.rs (309) — from_persistence_snapshot / persistence_snapshot.
- run_record.rs (116), limits.rs (150), admission.rs (73).

Move-only: bodies are unchanged; struct definitions with private fields stay in
mod.rs so descendant submodules keep field access, and only the handful of
cross-module `Inner`/`RunRecord` methods and free fns were widened from
module-private to `pub(super)` (still crate-internal — the engine is
`pub(crate)`). Public paths (`turn_state_engine::TurnStateEngine`,
`::TurnStateStoreLimits`, `::DEFAULT_RUNNER_LEASE_TTL_SECONDS`) are unchanged; no
file now exceeds 1,500 lines, so the `arch-exempt: large_file` header is dropped.

Also runs `cargo fmt -p ironclaw_turns`, cleaning formatting drift left by the
earlier import/perl edits in this branch (row_store*.rs and the migrated test
files). ironclaw_turns suite (15 binaries) + clippy --all-targets --all-features
green.
